moved wykop_client_wip.py to wykop_client.py -> added error handling in app.py changed variables.ini which now provides data for wykop_client.py
