+++ wykopproducer/variables.ini
@@ -5,6 +5,10 @@ RABBIT_PASSWORD = admin
 RABBIT_EXCHANGE = feed-exchange
 RABBIT_VHOST = PUDELEK
 RABBIT_QUEUE = pudelek-feed
+WYKOP_APP_KEY = '123'
+WYKOP_SECRET_KEY = '321'
+WYKOP_LOGIN = 'maasdasd'
+WYKOP_ACCOUNT_KEY = 'abcd'
 
 [DEV]
 RABBIT_HOST = rabbitmq
@@ -13,3 +17,7 @@ RABBIT_PASSWORD = admin
 RABBIT_EXCHANGE = feed-exchange
 RABBIT_VHOST = PUDELEK
 RABBIT_QUEUE = pudelek-feed
+WYKOP_APP_KEY = '123'
+WYKOP_SECRET_KEY = '321'
+WYKOP_LOGIN = 'maasdasd'
+WYKOP_ACCOUNT_KEY = 'abcd'
