+++ wykopproducer/wykop_client/wykop_client.py
@@ -1,77 +1,88 @@
 import hashlib
-import json
 
-import logger
+import requests
 
 WYKOP_ADD_ENTRY_URL = 'https://a2.wykop.pl/Entries/Entry/entry/'
 WYKOP_ADD_LINK_URL = 'https://a2.wykop.pl/Addlink/Add/'
 WYKOP_LINK_DRAFT_URL = 'https://a2.wykop.pl/Addlink/Draft/'
+WYKOP_LOG_IN_URL = 'https://a2.wykop.pl/Login/Index/'
+WYKOP_PREPARE_THUMBNAIL_URL = 'https://a2.wykop.pl/Addlink/Images/'
 
 
 class WykopClient:
-    def __init__(self, appkey, secret_key, user_key):
+    def __init__(self, appkey, secret_key, login, account_key):
         self.WYKOP_APP_KEY = appkey
         self.WYKOP_SECRET_KEY = secret_key
-        self.WYKOP_USER_KEY = user_key
+        # self.WYKOP_USER_KEY = user_key
+        self.WYKOP_USER_LOGIN = login
+        self.WYKOP_ACCOUNT_KEY = account_key
 
-    def prepare_link_draft(self, message):
-        link_draft_post_params = {'url': message['message']['link']}
-        link_draft_headers = {
-            'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
-                                               WYKOP_LINK_DRAFT_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
-                                               post_params=link_draft_post_params['url']),
-            'Content-type': 'application/x-www-form-urlencoded'
-        }
-        wykop_link_draft_client = self.requests.post(
-            self.WYKOP_SECRET_KEY,
+    def log_in(self):
+        log_in_postparams = {'login': self.WYKOP_USER_LOGIN,
+                             'accountkey': self.WYKOP_ACCOUNT_KEY}
+        log_in_headers = {'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
+                                                             WYKOP_LOG_IN_URL + 'appkey/' + self.WYKOP_APP_KEY + '/accountkey/' + self.WYKOP_ACCOUNT_KEY,
+                                                             post_params=','.join(
+                                                                 '{}'.format(log_in_postparams[key]) for key in
+                                                                 log_in_postparams))}
+        response = requests.post(
+            url=WYKOP_LOG_IN_URL + 'appkey/' + self.WYKOP_APP_KEY + '/accountkey/' + self.WYKOP_ACCOUNT_KEY,
+            data=log_in_postparams,
+            headers=log_in_headers)
+        return response.json()['data']['userkey']
+
+    def prepare_link_for_posting(self, message, userkey):
+        prepare_link_post_params = {'url': message['message']['link']}
+        prepare_link_headers = {'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
+                                                                   WYKOP_LINK_DRAFT_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + userkey,
+                                                                   prepare_link_post_params['url'])}
+        response = requests.post(
             WYKOP_LINK_DRAFT_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
-            data=link_draft_post_params,
-            headers=link_draft_headers
-        )
-        return json.loads(wykop_link_draft_client.json())
+            data=prepare_link_post_params,
+            headers=prepare_link_headers)
+        return response.json()
 
-    def add_link(self, message, draft_response):
-        key = ''  # patrz dokumentacja
-        addlink_post_params = {'title': message['message']['title'],
-                               'descritpion': message['message']['description'],
-                               'tags': ', '.join(message['message']['tags']),
-                               'photo': draft_response['data']['compact']['photos']['key'],
-                               'url': message['message']['link'],
-                               'plus18': True}
-        addlink_headers = {
-            'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
-                                               WYKOP_ADD_LINK_URL + 'key/' + key + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
-                                               post_params=','.join('{}'.format(addlink_post_params[key]) for key in
-                                                                    addlink_post_params)),
-            'Content-type': 'application/x-www-form-urlencoded'
-        }
-        wykop_addlink_client = self.requests.post(
-            WYKOP_ADD_LINK_URL + 'key/' + key + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
-            data=addlink_post_params, headers=addlink_headers)
-        return json.loads(wykop_addlink_client.json())
+    def prepare_news_thumbnail(self, prepare_link_response, userkey):
+        prepare_thumbnail_nameparam = prepare_link_response['data']['key']
+        prepare_thumbnail_headers = {'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
+                                                                        WYKOP_PREPARE_THUMBNAIL_URL + 'key/' + prepare_thumbnail_nameparam + '/appkey/' + self.WYKOP_APP_KEY + '/userkey/' + userkey)}
+        response = requests.post(
+            WYKOP_PREPARE_THUMBNAIL_URL + 'key/' + prepare_thumbnail_nameparam + '/appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
+            headers=prepare_thumbnail_headers)
+        return response.json()
 
-    def add_entry(self, addlink_response):
-        addentry_post_params = {
-            'body': 'Patrzcie co znalazlem ' + addlink_response['data']['compact']['full']['url'],
-            'embed': '',
-            'adultmedia': False
-        }
-        addentry_headers = {
-            'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
-                                               WYKOP_ADD_ENTRY_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
-                                               post_params=','.join(
-                                                   '{}'.format(addentry_post_params[key]) for key in
-                                                   addentry_post_params)),
-            'Content-type': 'application/x-www-form-urlencoded'
-        }
-        wykop_entry_client = self.requests.post(
-            WYKOP_ADD_ENTRY_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + self.WYKOP_USER_KEY,
-            data=addentry_post_params,
-            headers=addentry_headers)
-        if wykop_entry_client.status_code == 200:
+    def add_link_on_wykop(self, prepare_link_response, prepare_thumbnail_response, message, userkey):
+        add_link_nameparam = prepare_link_response['data']['key']
+        add_link_postparams = {'title': prepare_link_response['data']['title'],
+                               'description': message['message']['description'],
+                               'tags': ','.join(message['message']['tags']).replace(' ', ''),
+                               'photo': prepare_thumbnail_response['data'][0]['key'],
+                               'url': message['message']['link']}
+        add_link_headers = {'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
+                                                               WYKOP_ADD_LINK_URL + 'key/' + add_link_nameparam + '/appkey/' + self.WYKOP_APP_KEY + '/userkey/' + userkey,
+                                                               post_params=','.join(
+                                                                   '{}'.format(add_link_postparams[key]) for key in
+                                                                   add_link_postparams))}
+        response = requests.post(
+            WYKOP_ADD_LINK_URL + 'key/' + add_link_nameparam + '/appkey/' + self.WYKOP_APP_KEY + '/userkey/' + userkey,
+            data=add_link_postparams,
+            headers=add_link_headers)
+        return response.json()
+
+    def add_entry(self, addlink_response, userkey):
+        add_entry_postparams = {
+            'body': 'Sprawdź to https://www.wykop.pl/link/{}'.format(addlink_response['data']['id']),
+            'adultmedia': False}
+        add_entry_headers = {'apisign': self.create_md5checksum(self.WYKOP_SECRET_KEY,
+                                                                WYKOP_ADD_ENTRY_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + userkey,
+                                                                post_params=','.join(
+                                                                    '{}'.format(add_entry_postparams[key]) for key in
+                                                                    add_entry_postparams)),
+                             'Content-type': 'application/x-www-form-urlencoded'}
+        response = requests.post(WYKOP_ADD_ENTRY_URL + 'appkey/' + self.WYKOP_APP_KEY + '/userkey/' + userkey,
+                                 data=add_entry_postparams, headers=add_entry_headers)
+        if response.json()['data']['id']:
             return True
-        else:
-            return False, logger.info('Message has not been sent to wykop.pl')
 
     @staticmethod
     def create_md5checksum(secret, url, post_params=''):
@@ -83,10 +94,16 @@ class WykopClientMock:
     def __init__(self, *args, **kwargs):
         pass
 
-    def prepare_link_draft(self, *args, **kwargs):
+    def log_in(self):
+        return 'mocmock'
+
+    def prepare_link_for_posting(self, *args, **kwargs):
         return {'data': {'compact': {'photos': {'key': 'photokey'}}}}
 
-    def add_link(self, *args, **kwargs):
+    def prepare_news_thumbnail(self, *args, **kwargs):
+        return 'mockmock'
+
+    def add_link_on_wykop(self, *args, **kwargs):
         return {'data': {'compact': {'full': {'url': 'www.url.url'}}}}
 
     def add_entry(self, *args, **kwargs):
